Add support for GitHub releases
Add support for GitHub releases on the following architectures:

  * i386
  * amd64

Technically we should be able to support any architecture that has a
Linux port on it (or provides a SocketCAN interface). However, until
we can validate on arm/arm64, we won't release binaries unless
needed.

They're being built via the .goosarch file anyways, so building the
binaries isn't a major issue.