Add support for TCP transport
This adds support for TCP transport.

Unlike the multicast UDP connection, where each datagram contains the 30
bytes of data, the TCP stream uses 14-bytes, as it does not need to send
the Bus Identifier and Client Identifier for each frame. Rather, it is
configured initially when the TCP connection is established, and then
due to the point-to-point nature of TCP, it is implicitly known what bus
is sending what data.