Stop forwarding Tritium Settings packets
This also excludes the Tritium Settings packets from being forwarded
over SocketCAN. According to the Ethernet Interface Specification,
Settings packets are a setting for the bridge, and should not be bridged
onto the physical CAN network.

Technically, this shouldn't matter in the single bridge case, but if
there are multiple devices sending multicast UDP messages, it is
possible for us to receive a Settings packet that was sent from another
device, and attempt to forward it. This can happen if the setup is over
a router, or a configuration with multiple network interfaces.