Update README.md and Example 01 to use a static IP
Update the documentation to use a static IP for Example 01. This assumes
that you have assigned a static IP to the Tritium CAN-Ethernet Bridge
via the config tool, and then are either connected directly to the
Bridge, or have reserved the IP address on your router.

The reference to the dump command in the README has been removed, and
will probably not be implemented at this time. If you rely on this
behaviour, candump can be used on the SocketCAN interface.